`contact.repository` moved to contacts catalog
